+++ changelog.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- Calculate `L` in `cpw.characteristics`.
+
 ## v0.1.0 - 2023-01-11
 
 ### Added